+++ spring-cloud-tencent-plugin-starters/spring-cloud-starter-tencent-trace-plugin/src/main/java/com/tencent/cloud/plugin/trace/TraceClientPreEnhancedPlugin.java
@@ -1,0 +1,79 @@
+/*
+ * Tencent is pleased to support the open source community by making Spring Cloud Tencent available.
+ *
+ * Copyright (C) 2019 THL A29 Limited, a Tencent company. All rights reserved.
+ *
+ * Licensed under the BSD 3-Clause License (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://opensource.org/licenses/BSD-3-Clause
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed
+ * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ * CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ *
+ */
+
+package com.tencent.cloud.plugin.trace;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import com.tencent.cloud.plugin.trace.attribute.SpanAttributesProvider;
+import com.tencent.cloud.polaris.context.PolarisSDKContextManager;
+import com.tencent.cloud.rpc.enhancement.plugin.EnhancedPlugin;
+import com.tencent.cloud.rpc.enhancement.plugin.EnhancedPluginContext;
+import com.tencent.cloud.rpc.enhancement.plugin.EnhancedPluginType;
+import com.tencent.cloud.rpc.enhancement.plugin.PluginOrderConstant;
+import com.tencent.polaris.api.utils.CollectionUtils;
+import com.tencent.polaris.assembly.api.AssemblyAPI;
+import com.tencent.polaris.assembly.api.pojo.TraceAttributes;
+
+public class TraceClientPreEnhancedPlugin implements EnhancedPlugin {
+
+	private final PolarisSDKContextManager polarisSDKContextManager;
+
+	private final List<SpanAttributesProvider> spanAttributesProviderList;
+
+	public TraceClientPreEnhancedPlugin(PolarisSDKContextManager polarisSDKContextManager, List<SpanAttributesProvider> spanAttributesProviderList) {
+		this.polarisSDKContextManager = polarisSDKContextManager;
+		this.spanAttributesProviderList = spanAttributesProviderList;
+	}
+
+	@Override
+	public EnhancedPluginType getType() {
+		return EnhancedPluginType.Client.PRE;
+	}
+
+	@Override
+	public void run(EnhancedPluginContext context) throws Throwable {
+		Map<String, String> attributes = new HashMap<>();
+		if (CollectionUtils.isNotEmpty(spanAttributesProviderList)) {
+			for (SpanAttributesProvider spanAttributesProvider : spanAttributesProviderList) {
+				Map<String, String> additionalAttributes = spanAttributesProvider.getClientBaggageAttributes(context);
+				if (CollectionUtils.isNotEmpty(additionalAttributes)) {
+					attributes.putAll(additionalAttributes);
+				}
+			}
+		}
+
+		TraceAttributes traceAttributes = new TraceAttributes();
+		traceAttributes.setAttributes(attributes);
+		traceAttributes.setAttributeLocation(TraceAttributes.AttributeLocation.BAGGAGE);
+
+		AssemblyAPI assemblyAPI = polarisSDKContextManager.getAssemblyAPI();
+		assemblyAPI.updateTraceAttributes(traceAttributes);
+		Object otScope = traceAttributes.getOtScope();
+		if (otScope != null) {
+			context.getExtraData().put(SpanAttributesProvider.OT_SCOPE_KEY, otScope);
+		}
+	}
+
+	@Override
+	public int getOrder() {
+		return PluginOrderConstant.ClientPluginOrder.TRACE_CLIENT_PLUGIN_ORDER;
+	}
+}