+++ CHANGELOG.md
@@ -38,3 +38,4 @@
 - [feat:add admin http handler.](https://github.com/Tencent/spring-cloud-tencent/pull/1448)
 - [feat:support concurrency rate limit.](https://github.com/Tencent/spring-cloud-tencent/pull/1454)
 - [feat:support auth.](https://github.com/Tencent/spring-cloud-tencent/pull/1459)
+- [feat:upgrade trace plugin.](https://github.com/Tencent/spring-cloud-tencent/pull/1470)